+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- `notify-sdk` 子模块类优化注释、文档注释格式、代码换行格式。
+  - com.dwarfeng.notify.sdk.bean.entity.FastJsonDispatcherInfo。
+
 - `notify-stack` 子模块类优化注释、文档注释格式、代码换行格式。
   - com.dwarfeng.notify.stack.service.RouterInfoMaintainService。